Make error propagation consistent
Every error in the code now becomes an error event in the main thread
where possible. This, in turn, triggers a "crash" UI that shows a stack
trace. Errors that don't reflect programming bugs are caught and handled
separately so they don't propagate to here. Crashes are treated as
unrecoverable.