revive unit stress
I measure:

$ python3 unit-stress.py
8.05ms each; est 75 clients
8.12ms each; est 74 clients
7.78ms each; est 77 clients

I suspect most of this time is opus, but I have not verified that yet.

This is consistent with what I saw with full end to end stress testing, which could handle 50 clients but not 75. Of course, end to end testing has more overhead, but probably not much?