add full server-side support for shared memory sharding
I've tested this with two clients on jefftk.com, manually setting the API path, and it works.

Next step is to configure production and run load tests.

This requires a flag to enable, so it is not a problem to deploy in its current state and we can have both in parallel if we need to (though they won't share state, of course)

I'm expecting we will need a small client side change to choose which API to talk to