Fix monitoring
When we are connected, the user summary should never logically be empty, but it is actually empty right after we change positions.  I think this is because we don't have any audio yet, and position is sent with the summary and only when audio is sent.  This means we remove everyone from the mixing console and re-add, breaking monitoring.